+++ scripts/macros/spells/conjureElemental.js
@@ -23,7 +23,7 @@ export async function conjureElemental({speaker, actor, token, character, item, 
     };
     let animation = chris.getConfiguration(workflow.item, 'animation') ?? 'default';
     if (chris.jb2aCheck() != 'patreon' || !chris.aseCheck()) animation = 'none';
-    await summons.spawn(sourceActors, updates, 3600, workflow.item, true, undefined, undefined, workflow.token, animation);
+    await summons.spawn(sourceActors, updates, 3600, workflow.item, workflow.token, workflow.item.system?.range?.value, {'useActorOrigin': true, 'spawnAnimation': animation});
     let effect = chris.findEffect(workflow.actor, 'Concentrating');
     if (!effect) return;
     async function effectMacro () {