+++ scripts/module.js
@@ -132,7 +132,7 @@ Hooks.once('ready', async function() {
         Hooks.on('createChatMessage', addChatButton);
     }
     await loadTriggers();
-    //checkModules(); Want to test before releasing
+    checkModules();
     if (game.settings.get('chris-premades', 'Condition Resistance')) {
         Hooks.on('midi-qol.postPreambleComplete', macros.conditionResistanceEarly);
         Hooks.on('midi-qol.RollComplete', macros.conditionResistanceLate);